+++ src/mkvtoolnix-gui/info/initial_scan.cpp
@@ -1,5 +1,6 @@
 #include "common/common_pch.h"
 
+#include <QDebug>
 #include <QObject>
 
 #include "mkvtoolnix-gui/info/initial_scan.h"
@@ -7,8 +8,10 @@
 
 namespace mtx { namespace gui { namespace Info {
 
-InitialScan::InitialScan(Util::KaxInfo &info)
+InitialScan::InitialScan(Util::KaxInfo &info,
+                         Util::KaxInfo::ScanType type)
   : m_info{info}
+  , m_type{type}
 {
 }
 
@@ -17,7 +20,7 @@ InitialScan::~InitialScan() {
 
 void
 InitialScan::run() {
-  m_info.run();
+  m_info.runScan(m_type);
 }
 
 void