+++ src/mkvtoolnix-gui/info/initial_scan.h
@@ -10,9 +10,10 @@ namespace mtx { namespace gui { namespace Info {
 class InitialScan : public Util::Runnable {
 protected:
   Util::KaxInfo &m_info;
+  Util::KaxInfo::ScanType m_type;
 
 public:
-  explicit InitialScan(Util::KaxInfo &info);
+  explicit InitialScan(Util::KaxInfo &info, Util::KaxInfo::ScanType type);
   ~InitialScan();
 
   virtual void run() override;