+++ src/mkvtoolnix-gui/main_window/main_window.cpp
@@ -844,4 +844,14 @@ MainWindow::startStopQueueSpinner(bool start) {
   }
 }
 
+void
+MainWindow::startQueueSpinner() {
+  startStopQueueSpinner(true);
+}
+
+void
+MainWindow::stopQueueSpinner() {
+  startStopQueueSpinner(false);
+}
+
 }}